# Decode cache responses faster