I coded `scanGetResponseLine()` using `IndexByte()` instead of `SScanf()`.
Also added a benchmark.

```
name                   old time/op    new time/op    delta
ScanGetResponseLine-4    4.06µs ± 1%    0.13µs ± 4%  -96.89%  (p=0.008 n=5+5)

name                   old alloc/op   new alloc/op   delta
ScanGetResponseLine-4      128B ± 0%       24B ± 0%  -81.25%  (p=0.008 n=5+5)

name                   old allocs/op  new allocs/op  delta
ScanGetResponseLine-4      7.00 ± 0%      1.00 ± 0%  -85.71%  (p=0.008 n=5+5)
```

(Note I still copy the `line` to a `string`, because we need a string to put in `Item.Key`.  I'm copying more than is strictly needed but it simplifies the code)