@bboreham 2022-08-15
Upstream PR at https://github.com/bradfitz/gomemcache/pull/154